application/app_requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
numpy==2.1.3
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
psycopg2==2.9.10
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
